Move player class to a separate file
